chore: Move ConnectorSelectModal
