+++ src/components/DownloadModal/index.ts
@@ -2,5 +2,6 @@ export { InitialScreen } from "./InitialScreen";
 export { DownloadSettingsScreen } from "./DownloadSettingsScreen";
 export { DownloadTasksScreen } from "./DownloadTasksScreen";
 export { UploadTasksScreen } from "./UploadTasksScreen";
+export { ConnectorSelectionModal } from "./ConnectorSelectionModal";
 export * from "./types";
 export * from "./utils";
